+++ api/src/main/java/com/melodyguard/api/service/UserService.java
@@ -1,18 +1,23 @@
 package com.melodyguard.api.service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.melodyguard.api.dto.role.RoleRequest;
 import com.melodyguard.api.dto.user.UserRequest;
 import com.melodyguard.api.dto.user.UserResponse;
+import com.melodyguard.api.entity.Role;
 import com.melodyguard.api.entity.User;
 import com.melodyguard.api.exception.ElementNotFoundException;
+import com.melodyguard.api.exception.UnauthorizedAccessException;
 import com.melodyguard.api.mapper.UserMapper;
 import com.melodyguard.api.repository.UserRepository;
+import com.melodyguard.api.util.AuthorizationAccess;
 
 @Service
 public class UserService {
@@ -20,22 +25,41 @@ public class UserService {
     @Autowired UserMapper mapper;
 
     public List<UserResponse> getAllUsers(){
+        if (!AuthorizationAccess.hasRole(Role.ADMIN)) {
+            throw new UnauthorizedAccessException("Unauthorized Access, only admins can perform this action.");
+        }
+
         return repository.findAll().stream().map(user -> mapper.toDto(user)).collect(Collectors.toList());
     }
 
-    public UserResponse getUserById(String id){
-        return repository.findById(id).map(mapper::toDto).orElseThrow(() -> new ElementNotFoundException("User", id));
+    public User getUserById(String id){
+        if (!AuthorizationAccess.hasRole(Role.ADMIN)) {
+            throw new UnauthorizedAccessException("Unauthorized Access, only admins can perform this action.");
+        }
+
+        return repository.findById(id).orElseThrow(() -> new ElementNotFoundException("User", id));
     }
 
-    public UserResponse updateUserRoles(UserRequest userRequest){
-        
-        User updatedUser = repository.save(mapper.toEntity(userRequest));
+    public User updateUserRoles(String userId, RoleRequest roleRequest){
+        if (!AuthorizationAccess.hasRole(Role.ADMIN)) {
+            throw new UnauthorizedAccessException("Unauthorized Access, only admins can perform this action.");
+        }
 
-        return mapper.toDto(updatedUser);
+        Optional<User> userById = repository.findById(userId);
+
+        if (userById.isPresent()) {
+            User user = userById.get();
+            user.setRole(roleRequest.getRole());
+            return repository.save(user);
+        } else throw new ElementNotFoundException("User", userId);
     }
 
 
     public ResponseEntity<String> delete(String id){
+        if (!AuthorizationAccess.hasRole(Role.ADMIN)) {
+            throw new UnauthorizedAccessException("Unauthorized Access, only admins can perform this action.");
+        }
+
         getUserById(id); // fails when unvalid id
         repository.deleteById(id);
         return ResponseEntity.ok(String.format("User with Id `%s` deleted successfully.", id));
